changed contentLength() from int to long

+++ org.springframework.core/src/main/java/org/springframework/core/io/AbstractFileResolvingResource.java
@@ -129,7 +129,7 @@ public abstract class AbstractFileResolvingResource extends AbstractResource {
 	}
 
 	@Override
-	public int contentLength() throws IOException {
+	public long contentLength() throws IOException {
 		URL url = getURL();
 		if (ResourceUtils.isFileURL(url)) {
 			// Proceed with file system resolution...
